2주차 - 🚀 2단계 - 중복 제거와 구조화 (#26)
* docs: step2 진행을 위해 cursor 지침 문서 추가

* feat: 존재하지 않는 제품으로 대여 요청 시나리오 테스트 추가

* feat: 대여용이 아닌 제품으로 대여 요청 시나리오 테스트 추가

* feat: 재고가 부족한 제품으로 대여 요청 시나리오 테스트 추가

* feat: 존재하지 않는 예약으로 대여 요청 시나리오 테스트 추가

* feat: 예약 없이 대여 요청 (워크인 고객) 시나리오 테스트 추가

* docs: 엣지/예외 시나리오 feature 추가

* modify: 예외 메시지는 검증 대상에서 제한다.

* docs: cursor 가이드라인 구체화 (예외 메시지는 검증 대상에서 제외)

* docs: 시나리오를 문장형으로 작성하라

* fix: 두개의 테스트에서 변화를 주는 테스트 픽스처가 겹치지 않도록 하라

* fix: 예외 다루는 GlobalExceptionHandler 추가, 500 대신 BAD_REQUEST와 CONFLICT 로 세분화

* refactor: 워크인 고객에 대해 예약 고객과 동일한 Step definition 사용하도록 하라

* refactor: 재고가 부족한 케이스에 대해 step definition 공통화

* refactor: 존재하지 않는 상품에 대해step definition 공통화

* refactor: 존재하지 않는 예약 대해step definition 공통화

* chore: 미사용 코드 제거

* chore: 미사용 코드 제거

* refactor: 판매용 제품에 대해 step definition 공통화

* adjust: 재고부족 시나리오에서 '재고가 부족할때 대여한다' 를 when으로 표현하며 숫자는 시나리오에서 제거

* adjust: 시나리오는 추상화하고, 테스트에서 구체적 수치를 지정하라

---------

+++ src/test/java/com/camping/admin/helper/RequestFactory.java
@@ -9,7 +9,7 @@ public class RequestFactory {
 
     public static HashMap<String, Object> createRentalRequest(Long reservationId,
                                                               Long productId,
-                                                              Long quantity) {
+                                                              Integer quantity) {
         var request = new HashMap<String, Object>();
         request.put("reservationId", reservationId);
         request.put("productId", productId);
